LPD-34397 Rename and simplify
