+++ modules/apps/frontend-data-set/frontend-data-set-admin-web/src/main/java/com/liferay/frontend/data/set/admin/web/internal/model/listener/CompanyModelListener.java
@@ -5,7 +5,7 @@
 
 package com.liferay.frontend.data.set.admin.web.internal.model.listener;
 
-import com.liferay.frontend.data.set.admin.model.DataSetModelManager;
+import com.liferay.frontend.data.set.admin.manager.FDSAdminObjectDefinitionsManager;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.Company;
@@ -25,13 +25,14 @@ public class CompanyModelListener extends BaseModelListener<Company> {
 	public void onAfterCreate(Company company) throws ModelListenerException {
 		TransactionCommitCallbackUtil.registerCallback(
 			() -> {
-				_dataSetModelManager.checkCompany(company);
+				_fdsAdminObjectDefinitionsManager.generateObjectDefinitions(
+					company);
 
 				return null;
 			});
 	}
 
 	@Reference
-	private DataSetModelManager _dataSetModelManager;
+	private FDSAdminObjectDefinitionsManager _fdsAdminObjectDefinitionsManager;
 
 }
